requirements.txt
delegator.py
click
pyyaml
click_log
click_default_group
halo
crayons
